Minor modifications to poissonDist and binomialDist
